Damage Adjust
Adjusted Damage Script (now "TrapDamage" script) and removed unnecessary serialized fields. TrapDamage script now only handles traps dealing damage to the player, and has a bool to determine whether damage is dealt once to the player upon entering the trigger, or over time as long as the player is in the trigger.